Add ZXing core dependency
The Google ZXing core library has been added as a dependency to the `app` module.
This allows the application to utilize the core functionalities of the ZXing library for barcode and QR code processing.

Specifically, the following dependency was added to `app/build.gradle.kts`:
`implementation("com.google.zxing:core:3.5.3")`